Move o.a.c.net.MTOCLQ to o.a.c.concurrent and o.a.c.net.Crc to o.a.c.utils for general use
