+++ src/java/org/apache/cassandra/net/FrameDecoderLZ4.java
@@ -26,7 +26,7 @@ import io.netty.channel.ChannelPipeline;
 import net.jpountz.lz4.LZ4Factory;
 import net.jpountz.lz4.LZ4SafeDecompressor;
 
-import static org.apache.cassandra.net.Crc.*;
+import static org.apache.cassandra.utils.Crc.*;
 
 /**
  * Framing format that compresses payloads with LZ4, and protects integrity of data in movement with CRCs
