+++ src/java/org/apache/cassandra/net/InboundMessageHandlers.java
@@ -30,6 +30,7 @@ import io.netty.channel.Channel;
 import org.apache.cassandra.locator.InetAddressAndPort;
 import org.apache.cassandra.metrics.InternodeInboundMetrics;
 import org.apache.cassandra.net.Message.Header;
+import org.apache.cassandra.utils.Crc;
 
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 import static org.apache.cassandra.utils.MonotonicClock.Global.approxTime;
