+++ test/unit/org/apache/cassandra/concurrent/ManyToOneConcurrentLinkedQueueTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.cassandra.net;
+package org.apache.cassandra.concurrent;
 
 import java.util.BitSet;
 import java.util.NoSuchElementException;
